+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## IN PROGRESS
 
 * Leverage cookie-based authentication in all API requests. Fixes UILDP-77.
+* List all saved queries (not just the first ten), sorted alphabetically. Fixes UILDP-96.
 
 ## [1.10.0](https://github.com/folio-org/ui-ldp/tree/v1.10.0) (2023-02-24)